+++ crates/module-system/sov-modules-api/src/transaction/mod.rs
@@ -178,39 +178,6 @@ impl<R: TransactionCallable, S: Spec> PartialEq for Transaction<R, S> {
 }
 impl<R: TransactionCallable, S: Spec> Eq for Transaction<R, S> {}
 
-/// a [`Transaction`] with the `runtime_call` removed
-pub struct TransactionWithoutCall<S: Spec> {
-    /// The signature of the transaction.
-    pub signature: <S::CryptoSpec as CryptoSpec>::Signature,
-    /// The public key of the sender of the transaction.
-    pub pub_key: <S::CryptoSpec as CryptoSpec>::PublicKey,
-    /// The generation of the transaction (for uniqueness).
-    pub generation: u64,
-    /// The transaction metadata. Contains gas parameters and the chain ID.
-    pub details: TxDetails<S>,
-}
-
-impl<S: Spec> TransactionWithoutCall<S> {
-    /// Construct a [`Transaction`] by adding back the appropriate `CallMessage`.
-    pub fn with_call_v0<R: TransactionCallable>(self, runtime_call: R::Call) -> Transaction<R, S> {
-        let TransactionWithoutCall {
-            generation,
-            details,
-            signature,
-            pub_key,
-        } = self;
-        Transaction {
-            versioned_tx: VersionedTx::V0(Version0 {
-                signature,
-                pub_key,
-                runtime_call,
-                generation,
-                details,
-            }),
-        }
-    }
-}
-
 /// Errors that can be raised by the [`Transaction::verify`] method.
 #[derive(Error, Debug)]
 pub enum TransactionVerificationError<GU: Gas> {
@@ -300,17 +267,9 @@ impl<R: TransactionCallable, S: Spec> Transaction<R, S> {
     }
 
     /// Extract the runtime call from the transaction
-    pub fn split(self) -> (TransactionWithoutCall<S>, R::Call) {
+    pub fn call(self) -> R::Call {
         match self.versioned_tx {
-            VersionedTx::V0(inner) => (
-                TransactionWithoutCall {
-                    signature: inner.signature.clone(),
-                    pub_key: inner.pub_key.clone(),
-                    generation: inner.generation,
-                    details: inner.details.clone(),
-                },
-                inner.runtime_call.clone(),
-            ),
+            VersionedTx::V0(inner) => inner.runtime_call.clone(),
         }
     }